util.cpp: return double value for calc_phred_avg

+++ src/btllib/util.cpp
@@ -185,7 +185,7 @@ calc_phred_avg(const std::string& qual, const size_t start_pos, size_t len)
     phred_sum += delog_phred;
   }
 
-  return (uint32_t)(-10 * log10(phred_sum / len));
+  return -10 * log10(phred_sum / len);
 }
 
 void
